+++ ERP-Botequim_Mauro/database/migrations/2024_03_06_195314_create_paid__credits_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('Date');
 
             //*Inicio da chave estrangeira
-            $table->integer('Id_client')->unsigned();
+            $table->unsignedBigInteger('Id_client')->unsigned();
             $table->foreign('Id_client')->references('id')->on('clients');
 
             $table->timestamps();
