+++ ERP-Botequim_Mauro/database/migrations/2024_03_06_195415_create_stocks_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('Stock_code');
 
             //*Inicio da chave 
-            $table->integer('Id_product')->unsigned();
+            $table->unsignedBigInteger('Id_product')->unsigned();
             $table->foreign('Id_product')->references('id')->on('products');
 
             $table->timestamps();
