+++ ERP-Botequim_Mauro/database/migrations/2024_03_06_195443_create_sale__histories_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->string('Product_price');
 
             //*Inicio da chave estrangeira
-            $table->integer('Id_stock')->unsigned();
+            $table->unsignedBigInteger('Id_stock')->unsigned();
             $table->foreign('Id_stock')->references('id')->on('stocks');
 
             $table->timestamps();
